Finished

+++ Labb1/int_buffer.h
@@ -1,6 +1,6 @@
 // Labb1, Programmeringsmetodik
 // Oskar Rubensson (osru1900)
-// int_buffer.h, 2021-11-22 - 2021-11-22
+// int_buffer.h, 2021-11-10 - 2021-11-22
 // Contains the definition for int_buffer
 
 
@@ -18,14 +18,14 @@ class int_buffer {
     public:
     
         int_buffer() = delete;                              // Remove default-constructor
-        explicit int_buffer( size_t size );                 // Default construct
-        int_buffer( const int* source , size_t size );      // Resize construct
+        explicit int_buffer( size_t size );                 // Default construct (explicit)
+        int_buffer( const int* source , size_t size );      // Resize (copy) construct
         int_buffer( const int_buffer& rhs );                // Copy construct
         int_buffer( int_buffer&& rhs ) noexcept;            // Move construct
         int_buffer& operator=( const int_buffer& rhs );     // Copy assign
         int_buffer& operator=( int_buffer&& rhs ) noexcept; // Move assign
-        int& operator[]( size_t index );                    // Get at index
-        const int& operator[]( size_t index ) const;        // Get at index (const)
+        int& operator[]( int index );                       // Get at index
+        const int& operator[]( int index ) const;           // Get at index (const)
         [[nodiscard]]size_t size() const;                   // Get size
         void push_back(int value);                          // Push element at last-index
